requirements.txt
streamlit
tensorflow
scikit-learn
Pillow
numpy
